requirements.txt
flask
flask-cors
requests
python-dotenv
fpdf
google-generativeai
